Use Node 10 compatible code and a little refactoring
